+++ packages/shared/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @flowglad/shared
 
+## 0.10.5
+
+### Patch Changes
+
+- Rm list subscriptions
+
 ## 0.10.4
 
 ### Patch Changes